added `updateDataset(...)` to update metadata dataset
This new method make a request to the following address:

* PUT /api/datasets/:persistentId/versions/:draft?persistentId=$datasetId
* The body request should be a json with dataset metadata like as
  in addBasicDataset()

This Dataverse API endpoint let ordinary users to update or replace dataset
metadata. If the dataset is in a DRAFT state then the metadata will be
replaced, if the dataset already in a published state then a new DRAFT version
will be created.

See Dataverse API specs for details:

* https://guides.dataverse.org/en/latest/api/native-api.html#update-metadata-for-a-dataset